+++ src/main/java/racingcar/domain/RacingGame.java
@@ -1,4 +1,4 @@
-package racingcar.car;
+package racingcar.domain;
 
 import racingcar.strategy.MoveStrategy;
